requirements.txt
flask==3.0.0
python-telegram-bot==20.4
python-dotenv==1.0.0
werkzeug==3.0.1
